requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
Django==5.2.3
django-crispy-forms==2.4
django-extensions==4.1
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
whitenoise==6.9.0

# Morphometric Analysis Dependencies
cellpose[gui]
scikit-image
numpy
scipy
matplotlib
